local-code-agent/requirements.txt
google-generativeai
chromadb
sentence-transformers
tree-sitter==0.21.3
tree-sitter-languages
